style(rename): rename package name
- mango -> apple_mango
